+++ Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-﻿using System.Reflection;
+using System.Reflection;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("DrakeRenameit")]
 [assembly: AssemblyDescription("Manage your doors and chests better")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("DrakeMods")]
+[assembly: AssemblyCompany("DrakesWorkshop")]
 [assembly: AssemblyProduct("DrakeRenameit")]
 [assembly: AssemblyCopyright("Copyright ©  2025")]
 [assembly: AssemblyTrademark("")]
@@ -30,6 +30,4 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+// Assembly version is generated from mod.package.props (see obj/AssemblyVersion.g.cs).